correct some warnings

+++ resources.h
@@ -76,7 +76,7 @@ public:
     /* This function releases the resources in the input vector */
     void res_release(Wlk_Resources & occup) {
         /* Dimension check */
-        assert(available.size() == occup.ntypes());
+        assert((int)available.size() == occup.ntypes());
 
         for (int ii=0; ii < (int)available.size(); ii++){
             available[ii] += occup.get_resources(ii);
